+++ main.go
@@ -7,12 +7,13 @@ import (
   "github.com/hashicorp/packer-plugin-sdk/plugin"
 
   "github.com/mschuchard/packer-plugin-testinfra/version"
+  "github.com/mschuchard/packer-plugin-testinfra/provisioner"
 )
 
 func main() {
   // initialize packer plugin set for testinfra
   packerPluginSet := plugin.NewSet()
-  packerPluginSet.RegisterProvisioner(plugin.DEFAULT_NAME, new(TestinfraProvisioner))
+  packerPluginSet.RegisterProvisioner(plugin.DEFAULT_NAME, new(testinfra.Provisioner))
   packerPluginSet.SetVersion(version.PluginVersion)
 
   // execute packer plugin for testinfra