+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ### 1.2.1 (Next)
 - Pass marker and keyword expressions as singular strings to command executors.
+- Reorganize module into package.
 
 ### 1.2.0
 - Support executing tests by keyword substring expression.